refactor(auth): update response handling and improve error management in authentication flow

+++ src/main/java/com/example/test2/monitor/MonitorServiceImpl.java
@@ -1,11 +1,14 @@
 package com.example.test2.monitor;
 
+import com.example.test2.monitor.dto.MonitorDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -45,17 +48,46 @@ public class MonitorServiceImpl implements MonitorService {
     }
 
     @Override
-    public Monitor createMonitor(Monitor monitor) {
-        return monitorRepository.save(monitor);
+    public Monitor createMonitor(MonitorDTO monitor) {
+        var oldMonitor = monitorRepository.findMonitorByName(monitor.getName());
+        if (oldMonitor.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Monitor with name " + monitor.getName() + " already exists");
+        }
+
+        return monitorRepository.save(
+                Monitor
+                        .builder()
+                        .name(monitor.getName())
+                        .brand(monitor.getBrand())
+                        .price(monitor.getPrice())
+                        .build()
+        );
     }
 
     @Override
-    public Monitor updateMonitor(Monitor monitor) {
-        return monitorRepository.save(monitor);
+    public Monitor updateMonitor(int id, MonitorDTO monitor) {
+        var oldMonitor = monitorRepository.findMonitorByName(monitor.getName());
+        if (oldMonitor.isPresent() && oldMonitor.get().getId() != id) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Monitor with name " + monitor.getName() + " already exists");
+        }
+        var monitorToUpdate = monitorRepository.findById(id);
+        if (monitorToUpdate.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Monitor with id " + id + " not found");
+        }
+        Monitor updatedMonitor = monitorToUpdate.get();
+        updatedMonitor.setName(monitor.getName());
+        updatedMonitor.setBrand(monitor.getBrand());
+        updatedMonitor.setPrice(monitor.getPrice());
+
+        return monitorRepository.save(updatedMonitor);
     }
 
     @Override
     public void deleteMonitorById(int id) {
+        var monitor = monitorRepository.findById(id);
+        if (monitor.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Monitor with id " + id + " not found");
+        }
         monitorRepository.deleteById(id);
     }
